+++ ESLSCDLL/Board_ll.h
@@ -66,6 +66,7 @@ uint16_t* getVirtualDmaAddress(uint32_t drvno);
 uint32_t getDmaBufferSizeInBytes(uint32_t drvno);
 int64_t getCurrentInterruptCounter(uint32_t drvno);
 uint8_t WaitforTelapsed(int64_t microseconds);
+es_status_codes SaveMeasurementDataToFileBIN(const char* path, char* filename);
 #ifndef MINIMAL_BUILD
 void openFile(uint32_t drvno);
 void closeFile(uint32_t drvno);