+++ ESLSCDLL/linux/Board_ll.c
@@ -486,3 +486,7 @@ void initPerformanceCounter()
 	return;
 }
 
+es_status_codes SaveMeasurementDataToFileBIN(const char* path, char* filename)
+{
+	return es_no_error;
+}