+++ escam/lsc.cpp
@@ -499,17 +499,11 @@ es_status_codes Lsc::findCam(uint32_t drvno)
 }
 
 /**
- * @copydoc ExportMeasurementHDF5.
+ * @copydoc SaveMeasurementDataToFile.
  */
-es_status_codes Lsc::exportMeasurementHDF5(const char* path, char* filename)
+es_status_codes Lsc::SaveMeasurementDataToFile(const char* path, char* filename)
 {
-#ifdef WIN32
-	return DLLExportMeasurementHDF5(path, filename);
-#else
-	(void)path;
-	(void)filename;
-	return es_no_error;
-#endif
+	return DLLSaveMeasurementDataToFile(path, filename);
 }
 
 es_status_codes Lsc::waitForMeasureDone()